chore: avoid warnings
This is done through flags when no other option seems more sensible
(on generated ffi bindings and for repeating actions in tests).

The pci_bus_scan could be actually exposed in the Pci library if needed
in the future, I don't believe it's worth the effort to expose it if
nobody needs it for now.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>